test: fold parse SWAR tests into parse_test.go, rename swar_verify_test.go
- parse_swar_test.go's tests and the digit-run benchmark now live in
  parse_test.go next to the other parser tests, keeping the dropped
  IsIPv4/IsIPv6 SWAR experiment rationale as a section comment.
- swar_verify_test.go becomes swar_test.go. It stays in the root
  package rather than moving under swar/: it is an integration test of
  the swar consumers (EqualFold and the strings/bytes case wrappers,
  which sit above the swar package), and its ASCII fold references are
  shared with the fuzz targets in fuzz_test.go.
