+++ parse_test.go
@@ -2,10 +2,12 @@ package utils
 
 import (
 	"math"
+	"math/rand"
 	"strconv"
 	"strings"
 	"testing"
 
+	"github.com/gofiber/utils/v2/swar"
 	"github.com/stretchr/testify/require"
 )
 
@@ -914,4 +916,163 @@ func Test_ParseInt8_ParseUint8_EdgeCases(t *testing.T) {
 	require.Error(t, err)
 	_, err = ParseUint8("025x")
 	require.Error(t, err)
+}
+
+// The tests below cover the 8-digit SWAR parse step in ParseUint/ParseInt.
+//
+// A SWAR pre-validation for IsIPv4/IsIPv6 was prototyped alongside this work
+// and dropped: it lost on every shape (+10% to +49%) because the scalar
+// octet parse still has to run after the word checks on <= 15-byte inputs.
+// ParseUint won -16% to -27% at 8+ digits with a one-branch cost on 1-digit
+// inputs; the raw comparison lives in the introducing commit message.
+
+func Test_Parse8Digits_And_IsEightDigits(t *testing.T) {
+	t.Parallel()
+	for _, v := range []uint64{0, 1, 9, 10, 12345678, 99999999, 10000000, 90000009, 9999990} {
+		s := FormatUint(v)
+		for len(s) < 8 {
+			s = "0" + s
+		}
+		w := swar.Load8(s, 0)
+		require.True(t, isEightDigits(w), s)
+		require.Equal(t, v, parse8Digits(w), s)
+	}
+	// isEightDigits must reject every byte value outside '0'..'9' in every
+	// lane, including bytes >= 0xFA whose +6 carries into the next lane.
+	for c := range 256 {
+		lanes := []byte("11111111")
+		for lane := range 8 {
+			lanes[lane] = byte(c)
+			got := isEightDigits(swar.Load8(lanes, 0))
+			require.Equal(t, c >= '0' && c <= '9', got, "byte 0x%02x lane %d", c, lane)
+			lanes[lane] = '1'
+		}
+	}
+
+	// Pin parse8Digits itself on random 8-digit words, independent of the
+	// full-parser parity loop.
+	rng := rand.New(rand.NewSource(5)) //nolint:gosec // deterministic test data
+	for range 10000 {
+		v := uint64(rng.Intn(100000000))
+		s := FormatUint(v)
+		for len(s) < 8 {
+			s = "0" + s
+		}
+		w := swar.Load8(s, 0)
+		require.True(t, isEightDigits(w), s)
+		require.Equal(t, v, parse8Digits(w), s)
+	}
+}
+
+func Test_ParseSWAR_Parity(t *testing.T) {
+	t.Parallel()
+	rng := rand.New(rand.NewSource(3)) //nolint:gosec // deterministic test data
+	for range 200000 {
+		v := rng.Uint64() >> rng.Intn(60) // cover all digit counts
+		s := strconv.FormatUint(v, 10)
+		got, err := ParseUint(s)
+		require.NoError(t, err, s)
+		require.Equal(t, v, got, s)
+
+		gotI, err := ParseInt(s)
+		wantI, wantErr := strconv.ParseInt(s, 10, 64)
+		require.Equal(t, wantErr == nil, err == nil, s)
+		if wantErr == nil {
+			// On range errors strconv returns the clamped value while utils
+			// returns 0, so values are only comparable on success.
+			require.Equal(t, wantI, gotI, s)
+		}
+	}
+
+	// Error parity with strconv on syntax/overflow shapes that exercise the
+	// SWAR word steps, including the scalar handoff around 16-19 digits.
+	for _, s := range []string{
+		"", "a", "12345678a", "123456781234567812345", "18446744073709551615",
+		"18446744073709551616", "99999999999999999999", "00000000000000000001",
+		"1234567x", "12345678\n2345678", "184467440737095516150",
+		"0000000000000000000000042", "999999999999999999999999",
+	} {
+		want, wantErr := strconv.ParseUint(s, 10, 64)
+		got, gotErr := ParseUint(s)
+		require.Equal(t, wantErr == nil, gotErr == nil, "%q", s)
+		if wantErr == nil {
+			require.Equal(t, want, got, "%q", s)
+		} else {
+			require.Zero(t, got, "%q", s)
+			// The error must keep its strconv shape: *NumError carrying the
+			// function name and the ErrSyntax/ErrRange cause.
+			var numErr *strconv.NumError
+			require.ErrorAs(t, gotErr, &numErr, "%q", s)
+			require.Equal(t, "ParseUint", numErr.Func, "%q", s)
+			var wantNum *strconv.NumError
+			require.ErrorAs(t, wantErr, &wantNum)
+			require.ErrorIs(t, numErr.Err, wantNum.Err, "%q", s)
+		}
+	}
+}
+
+func Test_ParseInt_SWARPaths(t *testing.T) {
+	t.Parallel()
+	// Fast path: a non-digit inside the first word falls back to the scalar
+	// loop, which reports the syntax error.
+	for _, s := range []string{"1234x678", "12345678\x00", "123456789012345678x"} {
+		_, err := ParseInt(s)
+		require.Error(t, err, s)
+		want, wantErr := strconv.ParseInt(s, 10, 64)
+		require.Error(t, wantErr, s)
+		require.Equal(t, want, int64(0), s) // strconv returns 0 on syntax errors too
+	}
+
+	// Signed inputs with 8+ digit runs route through parseSigned's word path.
+	for _, s := range []string{
+		"-123456789", "+123456789012", "-9223372036854775808", "+9223372036854775807",
+		"-9223372036854775809", "-99999999999999999999", "-12345678a", "-000000001",
+	} {
+		got, gotErr := ParseInt(s)
+		want, wantErr := strconv.ParseInt(s, 10, 64)
+		require.Equal(t, wantErr == nil, gotErr == nil, s)
+		if wantErr == nil {
+			require.Equal(t, want, got, s)
+		} else {
+			require.Zero(t, got, s)
+		}
+	}
+
+	// The same dispatch feeds the narrower signed parsers.
+	v32, err := ParseInt32("-214748364")
+	require.NoError(t, err)
+	require.Equal(t, int32(-214748364), v32)
+	_, err = ParseInt32("-21474836480")
+	require.Error(t, err)
+
+	// 19-digit boundary on the sign-free fast path: MaxInt64 parses, one
+	// past it is a range error (not a syntax error), matching strconv.
+	maxV, err := ParseInt("9223372036854775807")
+	require.NoError(t, err)
+	require.Equal(t, int64(math.MaxInt64), maxV)
+	_, err = ParseInt("9223372036854775808")
+	var numErr *strconv.NumError
+	require.ErrorAs(t, err, &numErr)
+	require.ErrorIs(t, numErr.Err, strconv.ErrRange)
+}
+
+func Benchmark_ParseUint_DigitRuns(b *testing.B) {
+	nums := map[string]string{
+		"1digit":  "8",
+		"4digit":  "8080",
+		"6digit":  "123456",
+		"8digit":  "12345678",
+		"12digit": "123456789012",
+		"19digit": "1234567890123456789",
+	}
+	for name, s := range nums {
+		b.Run(name, func(b *testing.B) {
+			b.ReportAllocs()
+			for b.Loop() {
+				if _, err := ParseUint(s); err != nil {
+					b.Fatal("failed to parse uint")
+				}
+			}
+		})
+	}
 }
