+++ swar_test.go
